pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vibe-collab"
version = "0.9.7"
description = "AI-powered project collaboration framework with three-mode CLI (IDE/interactive/agent)"
readme = {file = "README.pypi.md", content-type = "text/markdown"}
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "VibeCollab Contributors" }
]
keywords = [
    "llm",
    "ai",
    "collaboration",
    "vibe-development",
    "project-management",
    "documentation",
    "yaml",
    "generator"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "PyYAML>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "Jinja2>=3.0",
]

[project.optional-dependencies]
llm = [
    "httpx>=0.25.0",
]
embedding = [
    "sentence-transformers>=2.0.0",
    "numpy>=1.20.0",
]
mcp = [
    "mcp>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/flashpoint493/VibeCollab"
Repository = "https://github.com/flashpoint493/VibeCollab"
Documentation = "https://github.com/flashpoint493/VibeCollab#readme"

[project.scripts]
vibecollab = "vibecollab.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/vibecollab"]
artifacts = ["src/vibecollab/patterns/**", "src/vibecollab/i18n/locales/**"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/schema",
    "/README.md",
    "/README.pypi.md",
    "/LICENSE",
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
